Revert "add message log to sharp"
