+++ DiscordBot/Models/Config.cs
@@ -7,5 +7,4 @@ internal class Config
 {
     public SocketTextChannel modChannel;
     public char prefix;
-    public bool messageLog;
 }
