+++ DiscordBot/Program.cs
@@ -16,6 +16,9 @@ class Program
     /// <param name="args">Commandline arguments.</param>
     static void Main(string[] args) => new Program().MainAsync().GetAwaiter().GetResult();
 
+    private DiscordSocketClient _client;
+    private CommandService _commands;
+    private GuildHandler _guildHandler;
     private IServiceProvider _services;
 
     /// <summary>
@@ -24,23 +27,23 @@ class Program
     /// <returns>A delayed task.</returns>
     public async Task MainAsync()
     {
-        DiscordSocketClient _client = new DiscordSocketClient();
-        CommandService _commands = new CommandService();
+        _client = new DiscordSocketClient();
+        _commands = new CommandService();
+
+        GuildService _guildService = new GuildService();
+        _guildHandler = new GuildHandler(_client, _guildService);
 
         _client.Log += Log;
+        _client.MessageReceived += HandleCommandAsync;
 
         _services = new ServiceCollection()
             .AddSingleton(_client)
             .AddSingleton(_commands)
-            .AddSingleton<StrikeHandler>()
+            .AddSingleton(_guildService)
             .AddSingleton<StrikeService>()
-            .AddSingleton<GuildHandler>()
-            .AddSingleton<MessageHandler>()
-            .AddSingleton<MessageService>()
-            .AddSingleton<GuildService>()
+            .AddSingleton<StrikeHandler>()
+            .AddSingleton(_guildHandler)
             .BuildServiceProvider();
-
-        var commandHandler = new CommandHandler(_client, _commands, _services.GetService<GuildHandler>(), _services.GetService<MessageHandler>(), _services);
 
         await _commands.AddModulesAsync(Assembly.GetEntryAssembly(), _services);
 
@@ -49,6 +52,37 @@ class Program
 
         // Block this task until the program is closed.
         await Task.Delay(-1);
+    }
+
+    /// <summary>
+    /// Asynchronously handles commands to the bot, which are SocketMessages.
+    /// </summary>
+    /// <param name="messageParam">The SocketMessage</param>
+    /// <returns>Is an async Task.</returns>
+    private async Task HandleCommandAsync(SocketMessage messageParam)
+    {
+
+        // Don't process the command if it was a system message
+        var message = messageParam as SocketUserMessage;
+        if (message == null) return;
+
+        // Create a WebSocket-based command context based on the message
+        var context = new SocketCommandContext(_client, message);
+
+        // Create a number to track where the prefix ends and the command begins
+        int argPos = 0;
+        // Determine if the message is a command based on the prefix and make sure no bots trigger commands
+        if (!(message.HasCharPrefix(_guildHandler.GetPrefix(context.Guild).Result, ref argPos) ||
+            message.HasMentionPrefix(_client.CurrentUser, ref argPos)) ||
+            message.Author.IsBot)
+            return;
+
+        // Execute the command with the command context we just
+        // created, along with the service provider for precondition checks.
+        await _commands.ExecuteAsync(
+            context: context,
+            argPos: argPos,
+            services: _services);
     }
 
     /// <summary>
